refactor(skillsselector): move logic into separate script for cleanliness
