+++ src/components/Form/Form.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect, useState } from "react";
-import { TextInput, Stack, ComboBox } from "@carbon/react";
-import { OnChangeData } from "@carbon/react/lib/components/ComboBox/ComboBox";
+import React from "react";
+import { TextInput, Stack } from "@carbon/react";
 import { SkillItem } from "../List/List";
+import SkillsSelector from "../SkillsSelector/SkillsSelector";
 
 interface UserFormProps {
   fullName: string;
@@ -16,21 +16,6 @@ interface UserFormProps {
   setIsCurrentClientInvalid: (isInvalid: boolean) => void;
 }
 
-function useDebounce(value: string, delay: number) {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-}
-
 const UserForm: React.FC<UserFormProps> = ({
   fullName,
   isFullNameInvalid,
@@ -42,9 +27,6 @@ const UserForm: React.FC<UserFormProps> = ({
   setSelectedSkillList,
   setIsCurrentClientInvalid,
 }) => {
-  const [comboValue, setComboValue] = useState("");
-  const debouncedSearchTerm = useDebounce(comboValue, 500);
-
   const handleFullNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { value } = e.target;
     setFullName(value);
@@ -58,45 +40,6 @@ const UserForm: React.FC<UserFormProps> = ({
     setCurrentClient(value);
     setIsCurrentClientInvalid(false);
   };
-
-  const handleSkillChange = ({ selectedItem }: OnChangeData<string>) => {
-    if (selectedItem) {
-      setSelectedSkillList((prevSkills) => {
-        // Avoid adding duplicates
-        if (prevSkills.find((skill) => skill.title === selectedItem)) {
-          return prevSkills;
-        }
-        const newSkill: SkillItem = {
-          id: prevSkills.length + 1,
-          title: selectedItem,
-          skillLevel: "Beginner", // Default skill level; adjust as needed
-        };
-        return [...prevSkills, newSkill];
-      });
-    }
-  };
-
-  const [skillList, setSkillList] = useState<string[]>([]);
-
-  const fetchSkills = async (input: string) => {
-    if (!input) {
-      setSkillList([]);
-      return;
-    }
-
-    const response = await fetch(`https://api.apilayer.com/skills?q=${input}`, {
-      method: "GET",
-      headers: {
-        apikey: "FrV9cdgbzq8Ff8As0G5mstPG1aYrF7Lg",
-      },
-    });
-    const result = await response.json();
-    setSkillList(result || []);
-  };
-
-  useEffect(() => {
-    fetchSkills(debouncedSearchTerm);
-  }, [debouncedSearchTerm]);
 
   return (
     <Stack gap={7}>
@@ -127,20 +70,7 @@ const UserForm: React.FC<UserFormProps> = ({
         autoComplete="organization"
         required
       />
-      <ComboBox
-        id={"skills-combo"}
-        items={skillList}
-        onChange={handleSkillChange}
-        aria-activedescendant="skills-combo"
-        onInputChange={(val: any) => {
-          const value =
-            typeof val === "string" ? val : (val?.target?.value ?? "");
-          setComboValue(value);
-        }}
-        titleText="Skills"
-        shouldFilterItem={() => true}
-        placeholder="e.g. React"
-      />
+      <SkillsSelector setSelectedSkillList={setSelectedSkillList} />
     </Stack>
   );
 };
